+++ lakeapi/orderbook.py
@@ -8,7 +8,7 @@ if TYPE_CHECKING:
 
 class OrderBookUpdater:
 	''' Maintains order book snapshot while iterating over a dataframe with order book deltas. '''
-	def __init__(self, df: 'pd.DataFrame', depth_limit_min: int = 3000, depth_limit_max: int = 5000):
+	def __init__(self, df: 'pd.DataFrame', depth_limit_min: int = 3000, depth_limit_max: int = 5000, debug: bool = False):
 		self.bid = Dict.empty(key_type = float64, value_type = float64)
 		self.ask = Dict.empty(key_type = float64, value_type = float64)
 		self.received_timestamp = None
@@ -16,8 +16,13 @@ class OrderBookUpdater:
 		self.depth_limit_min = depth_limit_min
 		self.depth_limit_max = depth_limit_max
 		self._bests_cache = List()
+		self._bests_cache.append(float('inf'))
 		self._bests_cache.append(0.)
-		self._bests_cache.append(0.)
+		# self._lazy_cache = List()
+		# self._lazy_cache.append(0.)
+		# self._lazy_cache.append(0.)
+
+		self._debug = debug
 		self.switch_to_next_day(df)
 
 	def switch_to_next_day(self, df: 'pd.DataFrame') -> None:
@@ -27,6 +32,29 @@ class OrderBookUpdater:
 		self._side_is_bid = df['side_is_bid'].astype('float64').values
 		self._price = df['price'].astype('float64').values
 		self._size = df['size'].astype('float64').values
+
+		if self.bid and self.ask:
+			self._check_bests()
+
+	def _check_bests(self) -> None:
+		i = 0
+		while True:
+			bbid, bask = self.get_bests()
+			if bbid >= bask:
+				if i == 0:
+					bid = sorted(self.bid.keys(), reverse=True)[:10]
+					ask = sorted(self.ask.keys())[:10]
+				# Some old price level was not removed from the book. We should remove it.
+				# We drop best bid and ask as we don't know which one is the stale one.
+				del self.bid[bbid]
+				del self.ask[bask]
+				self._bests_cache[0] = float('inf')
+				self._bests_cache[1] = 0.
+				i += 1
+			else:
+				if i != 0:
+					print('Warning: Best prices crossed, dropped', i, 'levels to fix it. Bid:', bid, 'Ask:', ask)
+				break
 
 	@staticmethod
 	@njit(cache = True)
@@ -40,7 +68,7 @@ class OrderBookUpdater:
 					if price in bid_book:
 						del bid_book[price]
 					if bests_cache[0] == price:
-						bests_cache[0] = 0.
+						bests_cache[0] = float('inf')
 				else:
 					bid_book[price] = size
 					if price > bests_cache[0]:
@@ -87,7 +115,7 @@ class OrderBookUpdater:
 		return self.current_index
 
 	def get_bests(self) -> Tuple[float, float]:
-		if not self._bests_cache[0]:
+		if self._bests_cache[0] == float('inf'):
 			self._bests_cache[0] = max(self.bid)
 		if not self._bests_cache[1]:
 			self._bests_cache[1] = min(self.ask)